Update for sprockets 4 and sassc-rails
The `sassc-ruby` script interface is different from `ruby-sass`, and at
the current time does not seem to support keyword arguments. Rework the
function to accept an option map instead.

The method of registering custom Sass functions is also different.
`Functions#declare` does not exist, and the custom functions must
be defined on specific modules provided by the processor library.
This is complicated by there being two sassc processors in general
use. `sprockets` provides `Sprockets::SasscProcessor` while
`sassc-rails` provides `SassC::Rails::SassTemplate` overriding the
processor that `sprockets` provides. Each of these processors has its
own module on which custom functions are defined, so we extract
the `svgo` function in to a module and include it into each.